UPDATE- SOLO SI HAY CAMBIOS

+++ lib/_backend.sh
@@ -153,7 +153,19 @@ backend_update() {
   sudo su - deploy <<EOF
   cd /home/deploy/${empresa_atualizar}
   pm2 stop ${empresa_atualizar}-backend
-  git pull
+
+  PULL_OUTPUT=\$(git pull)
+
+  echo "\$PULL_OUTPUT"
+
+  if echo "\$PULL_OUTPUT" | grep -q "Already up to date."; then
+    echo "✅ No hay cambios para actualizar. Saliendo..."
+    pm2 start ${empresa_atualizar}-backend
+    pm2 save
+    exit 0
+  fi
+
+  # Solo si hay cambios
   cd /home/deploy/${empresa_atualizar}/backend
   npm install --loglevel=error
   npm update -f
